% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_spp_poor_sites.R
\name{cull_spp_poor_sites}
\alias{cull_spp_poor_sites}
\title{Function to cull datasets thathave too few taxa}
\usage{
cull_spp_poor_sites(pol_dl, min.spp)
}
\arguments{
\item{pol_dl}{Neotoma download object}

\item{min.spp}{integer; minimum number of samples}
}
\value{
A Neotoma download object
}
\description{
This function identifies Neotoma dastasets which have at least a user-defined minimum number of taxa
}
\author{
M. Allison Stegner
}
