% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_coords}
\alias{extract_coords}
\title{Function to extract coordinates from a Neotoma download object}
\usage{
extract_coords(pol_dl_obj)
}
\arguments{
\item{pol_dl_obj}{Neotoma download object}
}
\value{
A three column matrix containing dataset ids, longitudes, and latitudes
}
\description{
This function extracts coordinates from a Neotoma download object
}
\author{
M. Allison Stegner
}
