% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dl.R
\name{map_dl}
\alias{map_dl}
\title{Map a Neotoma Download Obect}
\usage{
map_dl(tax_dec_dl, X, Y, add, color, label.sites, return.table)
}
\arguments{
\item{tax_dl_dl}{a Neotoma download object}
}
\value{
.table logical; should a matrix containing site name, dataset id, long, and lat be returned?

A four column matrix containing site name, dataset ids, longitudes, and latitudes
}
\description{
Function to map location of Neotoma sites
}
\author{
M. Allison Stegner
}
