% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_correlation.R
\name{pairwise_correlation}
\alias{pairwise_correlation}
\title{Function to compute pairwise correlations on a matrix
This function calculates pairwise spearmanns correlations on a data matrix and returns rhos and ps}
\usage{
pairwise_correlation(datamatx, method)
}
\arguments{
\item{datamatx}{is a data matrix of species (rows) abundance per site or level (cols)}
}
\value{
rhos A matrix of pairwise rhos

ps A matrix of pairwise p values
}
\description{
Function to compute pairwise correlations on a matrix
This function calculates pairwise spearmanns correlations on a data matrix and returns rhos and ps
}
\author{
M. Allison Stegner
}
