% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen_transform.R
\name{pollen_transform}
\alias{pollen_transform}
\title{Function to transform a pollen time series}
\usage{
pollen_transform(ts, method)
}
\arguments{
\item{ts}{a vector of pollen data}

\item{method}{method of transforming the time series. Options are asin.sqrt, log, logit, sqrt, or hellinger}
}
\value{
A transformed vector
}
\description{
This function to transforms a pollen time series using one of several methods
}
\author{
M. Allison Stegner
}
