% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_high_res.R
\name{select_high_res}
\alias{select_high_res}
\title{Function to compile sampling resolution information for a set of Neotoma sites}
\usage{
select_high_res(dl_obj, max.grain)
}
\arguments{
\item{dl_obj}{Neotoma download object}

\item{max.grain}{integer; The maximum allowable number of years represented per pollen sample}
}
\value{
A list object containing a vector of dataset ids for sites that have resolution less than max.grain and a matrix of dataset ids, durations, and resolutions
}
\description{
This function identifies total duration and number of samples/duration for a Neotoma download
}
\author{
M. Allison Stegner
}
