% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_min_duration.R
\name{ts_min_duration}
\alias{ts_min_duration}
\title{Function to identify datasets that are of sufficient duration}
\usage{
ts_min_duration(dl_obj, min.time, agemodels)
}
\arguments{
\item{dl_obj}{Neotoma download object}

\item{min.time}{numeric; minimum duration, in years. Sites with number of samples less than min.samples are excluded}

\item{agemodels}{list object of agemodels. Can be provided by user, or default age model from Neotoma is used.}
}
\value{
A list of Neotoma dataset ids for sites that have adequate duration.
}
\description{
This function identifies Neotoma dastasets which have at least a minimum amount of temporal coverage
}
\author{
M. Allison Stegner
}
