% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dl.R
\name{map_dl}
\alias{map_dl}
\title{Map a Neotoma Download Obect}
\usage{
map_dl(dl_obj, X, Y, add, col = NULL, label.sites = FALSE,
  return.table = TRUE)
}
\arguments{
\item{dl_obj}{a Neotoma download object}

\item{X}{numeric; longitude}

\item{Y}{numeric; latitude}

\item{add}{logical; should points be added to an existing map?}

\item{col}{what color should the points be?}

\item{label.sites}{logical; should dataset ids be added as text labels?}

\item{return.table}{logical; should a matrix containing site name, dataset id, long, and lat be returned?}
}
\value{
A four column matrix containing site name, dataset ids, longitudes, and latitudes
}
\description{
Function to map location of Neotoma sites
}
\author{
M. Allison Stegner
}
